Add spigot version to the config + usability improvements to the scripts.

+++ destroy.sh
@@ -3,6 +3,9 @@
 
 ./stop.sh
 
-echo "Destroying container."
-docker rm -fv $SPIGOT_NAME
-
+if ! docker ps -a|grep -q $SPIGOT_NAME; then
+  echo "Container does not exist. Skipping destroy."
+else
+  echo "Destroying container."
+  docker rm -fv $SPIGOT_NAME
+fi
